First commit of what will eventually morph into the core of this algorithm
Have a couple classes for a Graph and GraphPath that I'm going to ditch
likely in favor of just using networkx's Graph class. Initially thought
having something lightweight might be nice but just going to piggy back.
Starting an algorithm that doesn't work yet but will. saving edits for now